tested for reading function from folder to script
